networking: explain that user is responsible for non-default config
GCP has many ways to set up networking. The CSB is validated using one
configuration, with the expecation that it would work in all
configurations as long as the network connectivity is correct. The
intention of this text is to give users freedom to configure GCP as they
would like, while also giving them the resposibility of making sure that
it's right.

[#179942019](https://www.pivotaltracker.com/story/show/179942019)