docs: add some example machine types
The tier/machine type field is passed through to the Google API and any
value that the Google API accepts can be specified. Unfortunately the
Google docs are hard to get started with, so I've added some example
values that work today, so that users don't get blocked on this.

[#182125710](https://www.pivotaltracker.com/story/show/182125710)