update instruction
add postgres migration instructions for gcp. Customers who want to
migrate their  data from old gcp service broker instances to the new gcp
broker will want to follow the instructions outlined in the doc. The
migration is based on impoert / export feature of gcp cloudsql because
it provided the simplest experience as compared to disk based backups or
manual psql dumps. it doesn't require connecting directly to the DB with
psql or similar tools to run statements and the ownership of migrated
data is properly setup, which is a requirement for the new broker to be
able to create bindings after the import.